Fix several deprecated warnings

+++ src/main/java/es/udc/gac/bigdec/util/IOUtils.java
@@ -19,6 +19,7 @@
 package es.udc.gac.bigdec.util;
 
 import java.io.BufferedReader;
+import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.lang.reflect.InvocationTargetException;
@@ -27,6 +28,7 @@ import java.text.DecimalFormat;
 
 import org.apache.hadoop.fs.FSDataInputStream;
 import org.apache.hadoop.fs.FSDataOutputStream;
+import org.apache.hadoop.fs.FileStatus;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.io.compress.CompressionCodec;
@@ -46,7 +48,7 @@ public final class IOUtils {
 	private static final DecimalFormat ONE_DECIMAL_FORMAT = new DecimalFormat("0.0");
 	private static final DecimalFormat TWO_DECIMAL_FORMAT = new DecimalFormat("0.00");
 	private static final int ONE_MiB = 1024 * 1024;
-	private static final String APP_NAME_TAG = "["+RunEC.APP_NAME+"] ";
+	private static final String APP_NAME_TAG = "[" + RunEC.APP_NAME + "] ";
 	private static final String WARN_TAG = "[WARN] ";
 	private static final String ERROR_TAG = "[ERROR] ";
 	private static final String NEW_LINE = "\n";
@@ -55,7 +57,8 @@ public final class IOUtils {
 	private static Path OUTPUT_PATH2;
 	private static long splitSize;
 
-	private IOUtils() {}
+	private IOUtils() {
+	}
 
 	public static Path getOutputPath1() {
 		return OUTPUT_PATH1;
@@ -74,7 +77,7 @@ public final class IOUtils {
 	}
 
 	public static double ByteToMiB(long size) {
-		return size / ((double)ONE_MiB);
+		return size / ((double) ONE_MiB);
 	}
 
 	public static String formatOneDecimal(double number) {
@@ -98,7 +101,8 @@ public final class IOUtils {
 		System.exit(-1);
 	}
 
-	public static long getNumberOfSplits(org.apache.hadoop.conf.Configuration hadoopConfig, CLIOptions options, Configuration config, 
+	public static long getNumberOfSplits(org.apache.hadoop.conf.Configuration hadoopConfig, CLIOptions options,
+			Configuration config,
 			int parallelism) throws IOException {
 		Path inputPath1 = null, inputPath2 = null;
 		Path outputPath = null, mergeOutputPath = null;
@@ -122,12 +126,14 @@ public final class IOUtils {
 		isInputPath1Splitable = IOUtils.isInputPathSplitable(hadoopConfig, inputPath1);
 		blockSize = srcFS.getFileStatus(inputPath1).getBlockSize();
 		blockSizeMB = IOUtils.ByteToMiB(blockSize);
-		blocks = SingleEndSequenceInputFormat.getNumberOfSplits(inputPath1, inputPath1Length, isInputPath1Splitable, blockSize);
+		blocks = SingleEndSequenceInputFormat.getNumberOfSplits(inputPath1, inputPath1Length, isInputPath1Splitable,
+				blockSize);
 		splitsPerCore = options.getSplitsPerCore();
-		nsplits = parallelism*splitsPerCore;
+		nsplits = parallelism * splitsPerCore;
 		splitSize = 0;
 
-		logger.info("INIT: blocks {}, blockSize {}, nsplits {}, splitsPerCore {}, splitSize {}", blocks, blockSize, nsplits, splitsPerCore, splitSize);
+		logger.info("INIT: blocks {}, blockSize {}, nsplits {}, splitsPerCore {}, splitSize {}", blocks, blockSize,
+				nsplits, splitsPerCore, splitSize);
 
 		if (blocks >= nsplits) {
 			if (blocks > nsplits) {
@@ -152,7 +158,8 @@ public final class IOUtils {
 
 				if (splitSize < Configuration.MIN_SPLIT_SIZE) {
 					IOUtils.warn(String.format("the split size (%d bytes) is lower than "
-							+ "the minimum value. Using %d MiBytes instead", splitSize, Configuration.MIN_SPLIT_SIZE_MiB));
+							+ "the minimum value. Using %d MiBytes instead", splitSize,
+							Configuration.MIN_SPLIT_SIZE_MiB));
 					splitSize = Configuration.MIN_SPLIT_SIZE;
 					nsplits = IOUtils.divideRoundUp(inputPath1Length, splitSize);
 				}
@@ -166,7 +173,8 @@ public final class IOUtils {
 		hadoopConfig.setInt(Configuration.DFS_REPLICATION_KEY, config.HDFS_BLOCK_REPLICATION);
 		splitSizeMB = IOUtils.ByteToMiB(splitSize);
 
-		logger.info("END: blocks {}, blockSize {}, nsplits {}, splitsPerCore {}, splitSize {}", blocks, blockSize, nsplits, splitsPerCore, splitSize);
+		logger.info("END: blocks {}, blockSize {}, nsplits {}, splitsPerCore {}, splitSize {}", blocks, blockSize,
+				nsplits, splitsPerCore, splitSize);
 
 		if (options.getInputFile2() != null) {
 			inputPath2 = srcFS.resolvePath(new Path(options.getInputFile2()));
@@ -176,12 +184,15 @@ public final class IOUtils {
 			isInputPath2Splitable = IOUtils.isInputPathSplitable(hadoopConfig, inputPath2);
 
 			// Sanity checks
-			if ((isInputPath1Compressed && !isInputPath2Compressed) || (!isInputPath1Compressed && isInputPath2Compressed))
+			if ((isInputPath1Compressed && !isInputPath2Compressed)
+					|| (!isInputPath1Compressed && isInputPath2Compressed))
 				throw new UnsupportedOperationException("Both input files must be compressed or not compressed");
 
 			if (isInputPath1Compressed && isInputPath2Compressed) {
-				if ((isInputPath1Splitable && !isInputPath2Splitable) || (!isInputPath1Splitable && isInputPath2Splitable))
-					throw new UnsupportedOperationException("Both compressed input files must be splitable or not splitable");
+				if ((isInputPath1Splitable && !isInputPath2Splitable)
+						|| (!isInputPath1Splitable && isInputPath2Splitable))
+					throw new UnsupportedOperationException(
+							"Both compressed input files must be splitable or not splitable");
 			}
 
 			if (!isInputPath1Compressed && !isInputPath2Compressed) {
@@ -197,10 +208,13 @@ public final class IOUtils {
 			basePath = new Path(options.getOutputDir());
 
 		// Create output path if it does not exist
-		if (srcFS.exists(basePath)) {
-			if (!srcFS.isDirectory(basePath)) 
-				throw new RuntimeException("Output path is invalid: "+basePath+" is not a directory");
-		} else {
+		try {
+			FileStatus status = srcFS.getFileStatus(basePath);
+
+			if (!status.isDirectory()) {
+				throw new RuntimeException("Output path is invalid: " + basePath + " is not a directory");
+			}
+		} catch (FileNotFoundException e) {
 			srcFS.mkdirs(basePath);
 		}
 
@@ -209,9 +223,9 @@ public final class IOUtils {
 		if (basePath.isRoot())
 			concatPath = concatPath.substring(0, concatPath.length() - 1);
 
-		outputPath = new Path(concatPath+Configuration.SLASH+RunEC.APP_NAME+"-Output");
-		OUTPUT_PATH1 = new Path(outputPath+Configuration.SLASH+"output1");
-		OUTPUT_PATH2 = new Path(outputPath+Configuration.SLASH+"output2");
+		outputPath = new Path(concatPath + Configuration.SLASH + RunEC.APP_NAME + "-Output");
+		OUTPUT_PATH1 = new Path(outputPath + Configuration.SLASH + "output1");
+		OUTPUT_PATH2 = new Path(outputPath + Configuration.SLASH + "output2");
 		options.setOutputDir(outputPath.toString());
 
 		if (srcFS.exists(outputPath))
@@ -229,10 +243,14 @@ public final class IOUtils {
 			dstFS = mergeOutputPath.getFileSystem(hadoopConfig);
 
 			// Create merge output path if it does not exist
-			if (dstFS.exists(mergeOutputPath)) {
-				if(!dstFS.isDirectory(mergeOutputPath))
-					throw new RuntimeException("Merge output path is invalid: "+mergeOutputPath+" is not a directory");
-			} else {
+			try {
+				FileStatus status = dstFS.getFileStatus(mergeOutputPath);
+
+				if (!status.isDirectory()) {
+					throw new RuntimeException(
+							"Merge output path is invalid: " + mergeOutputPath + " is not a directory");
+				}
+			} catch (FileNotFoundException e) {
 				dstFS.mkdirs(mergeOutputPath);
 			}
 
@@ -242,42 +260,45 @@ public final class IOUtils {
 			options.setMergeOutputDir(outputPath.toString());
 		}
 
-		IOUtils.info("HDFS block replication = "+config.HDFS_BLOCK_REPLICATION);
-		IOUtils.info("HDFS block size = "+IOUtils.formatOneDecimal(blockSizeMB)+" MiBytes");
+		IOUtils.info("HDFS block replication = " + config.HDFS_BLOCK_REPLICATION);
+		IOUtils.info("HDFS block size = " + IOUtils.formatOneDecimal(blockSizeMB) + " MiBytes");
 
 		if (inputPath1LengthMB > 1)
-			IOUtils.info("input file = "+inputPath1+" ("+IOUtils.formatOneDecimal(inputPath1LengthMB)+" MiBytes)");
+			IOUtils.info(
+					"input file = " + inputPath1 + " (" + IOUtils.formatOneDecimal(inputPath1LengthMB) + " MiBytes)");
 		else
-			IOUtils.info("input file = "+inputPath1+" ("+inputPath1Length+" bytes)");
+			IOUtils.info("input file = " + inputPath1 + " (" + inputPath1Length + " bytes)");
 
 		if (inputPath2 != null) {
 			if (inputPath2LengthMB > 1)
-				IOUtils.info("input file = "+inputPath2+" ("+IOUtils.formatOneDecimal(inputPath2LengthMB)+" MiBytes)");
+				IOUtils.info("input file = " + inputPath2 + " (" + IOUtils.formatOneDecimal(inputPath2LengthMB)
+						+ " MiBytes)");
 			else
-				IOUtils.info("input file = "+inputPath2+" ("+inputPath2Length+" bytes)");
+				IOUtils.info("input file = " + inputPath2 + " (" + inputPath2Length + " bytes)");
 		}
 
-		IOUtils.info("output directory = "+outputPath);
-		if (mergeOutputPath != null) 
-			IOUtils.info("merge output directory = "+mergeOutputPath);
-		IOUtils.info("HDFS blocks = "+blocks);
-		IOUtils.info("parallelism = "+parallelism+" cores");
-		IOUtils.info("splits per core = "+splitsPerCore);
-		IOUtils.info("splits = "+nsplits);
+		IOUtils.info("output directory = " + outputPath);
+		if (mergeOutputPath != null)
+			IOUtils.info("merge output directory = " + mergeOutputPath);
+		IOUtils.info("HDFS blocks = " + blocks);
+		IOUtils.info("parallelism = " + parallelism + " cores");
+		IOUtils.info("splits per core = " + splitsPerCore);
+		IOUtils.info("splits = " + nsplits);
 		if (splitSizeMB >= 1)
-			IOUtils.info("split size = "+IOUtils.formatOneDecimal(splitSizeMB)+" MiBytes");
+			IOUtils.info("split size = " + IOUtils.formatOneDecimal(splitSizeMB) + " MiBytes");
 		else
-			IOUtils.info("split size = "+splitSize+" bytes");
+			IOUtils.info("split size = " + splitSize + " bytes");
 
 		return nsplits;
 	}
 
-	public static FileFormat getInputFileFormat(org.apache.hadoop.conf.Configuration hadoopConf, Path inputPath) throws IOException {
+	public static FileFormat getInputFileFormat(org.apache.hadoop.conf.Configuration hadoopConf, Path inputPath)
+			throws IOException {
 		FileSystem fs = FileSystem.newInstance(hadoopConf);
 		FileFormat format = FileFormat.FILE_FORMAT_UNKNOWN;
 
 		// Try to autodetect the input file format
-		IOUtils.info("detecting input file format from "+inputPath);
+		IOUtils.info("detecting input file format from " + inputPath);
 
 		if (IOUtils.isPathCompressed(hadoopConf, inputPath)) {
 			IOUtils.info("input file is compressed, assuming FASTQ format");
@@ -297,7 +318,8 @@ public final class IOUtils {
 		return format;
 	}
 
-	public static long writeHistogram(FileSystem fs, Path outputFile, int[] histogram, int offset, int length) throws IOException {
+	public static long writeHistogram(FileSystem fs, Path outputFile, int[] histogram, int offset, int length)
+			throws IOException {
 		FSDataOutputStream out = null;
 		long total = 0;
 		StringBuilder sb = new StringBuilder(length);
@@ -310,12 +332,12 @@ public final class IOUtils {
 				return total;
 			}
 
-			IOUtils.info("writing histogram to "+outputFile);
+			IOUtils.info("writing histogram to " + outputFile);
 
 			out = fs.create(outputFile);
 
 			for (int i = offset; i < length; i++) {
-				sb.append(i+SEP_CHAR).append(String.valueOf(histogram[i])).append(NEW_LINE);
+				sb.append(i + SEP_CHAR).append(String.valueOf(histogram[i])).append(NEW_LINE);
 				out.write(sb.toString().getBytes(StandardCharsets.UTF_8));
 				total += histogram[i];
 				sb.setLength(0);
@@ -332,21 +354,20 @@ public final class IOUtils {
 		BufferedReader br = null;
 		FSDataInputStream dis = null;
 		String line;
-		Path file;
+		Path file = null;
 		int[] histogram = new int[size];
 		int i = 0;
 
 		try {
-			if (!fs.exists(path))
-				IOUtils.error("Histogram file not found: "+path);
+			FileStatus status = fs.getFileStatus(path);
 
-			if (fs.isDirectory(path))
-				file = new Path(path+Configuration.SLASH+fs.listStatus(path)[0].getPath().getName());				
+			if (status.isDirectory())
+				file = new Path(path + Configuration.SLASH + fs.listStatus(path)[0].getPath().getName());
 			else
 				file = path;
 
 			if (!fs.exists(file))
-				IOUtils.error("Histogram file file not found: "+file);
+				IOUtils.error("Histogram file file not found: " + file);
 
 			logger.info("loading histogram from {}", file);
 
@@ -360,6 +381,8 @@ public final class IOUtils {
 
 				line = br.readLine();
 			}
+		} catch (FileNotFoundException e) {
+			IOUtils.error("Histogram file not found: " + path);
 		} finally {
 			if (br != null)
 				br.close();
@@ -367,7 +390,7 @@ public final class IOUtils {
 				dis.close();
 		}
 
-		if (delete)
+		if (delete && file != null)
 			fs.delete(file, true);
 
 		return histogram;
@@ -406,12 +429,14 @@ public final class IOUtils {
 		return sb.toString().length() + 15;
 	}
 
-	public static SingleEndSequenceInputFormat getInputFormatInstance(Class<? extends SingleEndSequenceInputFormat> inputFormatClass) {
+	public static SingleEndSequenceInputFormat getInputFormatInstance(
+			Class<? extends SingleEndSequenceInputFormat> inputFormatClass) {
 		Object obj = null;
 
 		try {
 			obj = inputFormatClass.getConstructor().newInstance();
-		} catch (NoSuchMethodException | SecurityException | InstantiationException | IllegalAccessException | IllegalArgumentException
+		} catch (NoSuchMethodException | SecurityException | InstantiationException | IllegalAccessException
+				| IllegalArgumentException
 				| InvocationTargetException e) {
 			IOUtils.error(e.getMessage());
 		}
